Update README.md
We had a discussion with PG where it was mentioned that blobfuse shouldn't be used for concurrent write operations on same file. But, it can be used for concurrent write operations to different files in the same instance.

Hence, updating the description for the --max-concurrency option for the blobfuse. 

Adding this description will help customer to understand for what purpose they can use the --max-concurrency option for.