Remove the maps_full_membuf() assert from write_callback().
The assert was wrong as we can have a smaller bc referring to a bigger membuf, in case
some read thread did get() before the writer and then writer got a small bc referring to
the smae membuf. Thouguh we write the entire membuf but the assert was claiming that
bc should refer to the entire membuf which may not be the case, as explained above.