Plug a race where readdirectory_cache::clear() may not delete the inode cache.
as the ref count was 2 in the above loop but by the time we acquired the
inode_map_lock below some other thread dropped one ref and now the following
vector loop was to drop the last ref. Since it was n't using decref() it was
nto dropping the caches correctly.