Remove wait-for-update logic
Currently we implement RMW by having the writeer wait for the membuf to be made
uptodate by the reader. The writer indefinitely waits for that. It may happen that
reader never markes the membuf uptodate as it's not able to read anything (read
0 bytes and eof set). In such case writer is left waiting forever.
We remove the indefinite wait logic and the writer jsut allows the reader to
proceed by releasing the lock and then acquires the lock to see if the membuf is
marked uptodate. if read went fine it should be marked uptodate. This is the common
case. But in case reader asked for a large membuf w/o checking the file size, and
hence read completed with read_bytes=0 and eof=1, in those rare cases, writer will
see that the membuf is still not uptodate after reader released the lock.
In such case the writer will retry the process of getting the bc vec and copying.