Undo reserved space and sync jobs added in case of failed RV state changes
JoinMV sets reserved space on RV. This must be undone if the fix-mv
workflow doesn't complete. We do this when refreshFromClusterMap() finds
out some RV in outofsync state in rvInfo but it's offline in clustermap.

Similarly there are other state transitions, f.e., StartSync adds sync
job to mvInfo, but if sync wokflow cannot complete and this RV is
reverted back to outofsync state by refreshFromClusterMap(), we must
purge the sync jobs.