Do not delete active MVs on node restart.
Previous implementation used to play safe and used to delete all MVs
from the local RVs, after waiting for the RVs to be maked offline.
Though this works, but this doesn't handle occassional node restarts
well. f.e., we can have the blobfuse process crash and come back up
again almost immediately but we would force all the node's RVs to be
purged, which will cause unnecessary data movement and also may result
in cluster stability issues if multiple processes crash simultaneously.

What this change does is better, it doesn't delete the MVs for which
these RVs are not marked offline/inband-offline.
Note that if an MV is not written to while the node was down, we can
safely reuse the MV's data. If it was written to, then the component
RV would be marked offline/inband-offline and such MVs can be deleted.