Handler must refresh clustermap before failing to client
This is important as we don't want the client to get stuck in
cm.RefreshClusterMap() waiting for an epoch refresh where there isn't
any as the clustermap has not changed.
Other than this, this change adds more resilience by having the client
retry for a fixed times on getting the NeedToRefreshClusterMap error
from the RPC server. This handles various transient errors that the
server may encounter.