Allow more MVs/RV during fix-mv workflow
We need this to allow one or more nodes going down and rebalancing those
MVs to other available nodes.
To keep the code simple we let config.MVsPerRV represent the MVsPerRV
for the new MV workflow and define cm.MVsPerRVScaleFactor which decides
how many more MVsPerRV we allow during fix-mv.