Update MV placer to honor fault domains when choosing component RVs
Fault domain may or may not be available for an RV. If available we
honor it and do not pick multiple component RVs of an MV from the same
fault domain. If fault domain is not available for an RV it can be used
for pplacing any MV. This way we work well w/ and w/o fault domain info
for RVs.
This change itself doesn't add support to query fault domain from IMDS
endpoint.

This also updates the placer so that it tries to distribute MVs
uniformly across various RVs picking RVs which are free over RVs which
are fuller.