Introduce handler.refreshFromClustermapAll() to refresh all MVs
If a node has say 10 RVs and 100 MVs on each RV and that node goes down
and comes back up. Now there's another node which is used by MVs as one
of the replicas along with the node that went down and came back up.
When those degraded MVs are fixed and we sent UpdateMV request to the
second node, it'll cause a refresh clustermap for all the MVs and for
each of them it'll have to fetch the clustermap which is a costly
operation. Instead we can fetch clustermap once and refresh all the
mvInfo for all MVs on all RVs on this node.