Turn off kernel attribute caching regardless of direct_io mount option
Since a new file starts with size -1 which is later updated to the
correct size after the write is done and the file is closed, but we
dont't invalidsate the kernel cache so kernel still reports -1 as the
file size. For now I'm turning off the kernel attribute cache fully,
later if we want to enable the kernel attribute cache we need to
invalidte the cache when file size changes.