Remove agents folder
we don't need the warmup agent as we do it inline.
We are now moving to a model where cache granularity is a chunk and not
file. This simplies the cache warmup case also as we don't need to worry
about cases where the full file was not read from Azure. Whatever chunks
of the file were read, they will be cached.